refactor: merge "project.lua" into "init.lua"
